requirements.txt
Django>=5.0,<6.0
supabase>=2.0,<3.0
python-dotenv>=1.0,<2.0
